+++ frontend/src/pages/dating/interest.tsx
@@ -1,32 +1,13 @@
-import AppBody from "../../components/share/app/AppBody"
-import YouAreMatchWithWhiteImg from "../../components/dating/pic/youarematchwithwhite.png"
-import YouAreMatchWithBlackImg from "../../components/dating/pic/youarematchwithblack.png"
-import HeartCheckingWhiteImg from "../../components/dating/pic/heartcheckingwhite.png"
-import HeartCheckingBlackImg from "../../components/dating/pic/heartcheckingblack.png"
-import ActivityPollWhiteImg from "../../components/dating/pic/activitypollwhite.png"
-import ActivityPollBlackImg from "../../components/dating/pic/activitypollblack.png"
-import RandomizationWhiteImg from "../../components/dating/pic/randomizationwhite.png"
-import RandomizationBlackImg from "../../components/dating/pic/randomizationblack.png"
-import TagOfInterestWhiteImg from "../../components/dating/pic/tagofinterestwhite.png"
-import TagOfInterestBlackImg from "../../components/dating/pic/tagofinterestblack.png"
-import DatingOptionsWhiteImg from "../../components/dating/pic/datingoptionwhite.png"
-import DatingOptionsBlackImg from "../../components/dating/pic/datingoptionblack.png"
-import DatingTutorialWhiteImg from "../../components/dating/pic/datingtutorialwhite.png"
-import DatingTutorialBlackImg from "../../components/dating/pic/datingtutorialblack.png"
 import { INTERESTS } from "../../components/dating/shared/interests"
-import { Heading, useBreakpointValue, Text, Box, Button, Grid, GridItem, Input, CheckboxGroup } from "@chakra-ui/react"
+import { Heading, Text, Box, Button, Grid, GridItem, Input, CheckboxGroup } from "@chakra-ui/react"
 import React, { useState } from "react"
 import DatingTag from "../../components/dating/DatingTag"
+import DatingAppBody from "../../components/dating/DatingAppBody"
 
 const TagOfInterest = () => {
     const [interests] = useState(INTERESTS)
     const [numOfInterest, setNumOfInterest] = useState(0)
     const [selectedInterests, setSelectedInterest] = useState<String[] | String>([])
-
-    const isMobile = useBreakpointValue({
-        base: false,
-        md: true,
-    })
 
     function handleSearch(event: React.KeyboardEvent<HTMLInputElement>) {
         var searchQuery = (document.getElementById("search") as HTMLInputElement).value
@@ -62,51 +43,7 @@ const TagOfInterest = () => {
     }
 
     return (
-        <AppBody
-            secondarynav={[
-                {
-                    name: "Randomization",
-                    to: "/dating",
-                    Icon: isMobile ? RandomizationWhiteImg : RandomizationBlackImg,
-                },
-                {
-                    name: "Heart checking",
-                    to: "/dating/likedyou",
-                    Icon: isMobile ? HeartCheckingWhiteImg : HeartCheckingBlackImg,
-                },
-                {
-                    name: "You are match with",
-                    to: "/dating/match",
-                    Icon: isMobile ? YouAreMatchWithWhiteImg : YouAreMatchWithBlackImg,
-                },
-                {
-                    name: "Activity poll",
-                    to: "/dating/poll",
-                    Icon: isMobile ? ActivityPollWhiteImg : ActivityPollBlackImg,
-                },
-                {
-                    name: "Tag of interest",
-                    to: "/dating/interest",
-                    isRight: true,
-                    disableText: true,
-                    Icon: isMobile ? TagOfInterestWhiteImg : TagOfInterestBlackImg,
-                },
-                {
-                    name: "Option",
-                    to: "/dating/option",
-                    isRight: true,
-                    disableText: true,
-                    Icon: isMobile ? DatingOptionsWhiteImg : DatingOptionsBlackImg,
-                },
-                {
-                    name: "Tutorial",
-                    to: "/dating/tutorial/welcome",
-                    isRight: true,
-                    disableText: true,
-                    Icon: isMobile ? DatingTutorialWhiteImg : DatingTutorialBlackImg,
-                },
-            ]}
-        >
+        <DatingAppBody>
             <Grid
                 templateAreas={`"topic button" "desc desc"`}
                 gridTemplateRows={"50px 75px"}
@@ -162,7 +99,7 @@ const TagOfInterest = () => {
                           <DatingTag key={interestId} bool={false} interestId={interestId} interestName={interestName} handleTag={handleTag} />
                       ))}
             </CheckboxGroup>
-        </AppBody>
+        </DatingAppBody>
     )
 }
